requirements.txt
# Core packages
numpy>=1.20.0
pandas>=1.3.0
pyyaml>=6.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# NLP and Deep Learning
transformers>=4.18.0
torch>=1.10.0
datasets>=2.14.0
tokenizers>=0.14.0
sentencepiece>=0.1.99
nltk>=3.6.0
spacy>=3.2.0

# Medical NLP specific
# Note: scispacy might need separate installation
# scispacy>=0.5.3
# negspacy>=1.0.3

# AWS Integration
boto3>=1.20.0
sagemaker>=2.70.0

# API and Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Utilities
tqdm>=4.62.0
pytest>=6.2.0
flake8>=4.0.0
black>=21.12b0
isort>=5.12.0
jupyter>=1.0.0 